be/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
groq==0.4.1
gTTS==2.4.0
langchain==0.1.0
langchain-groq==0.0.1
chromadb==0.4.18
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1

